Refactor files and clean function calls
