+++ config.json
@@ -1,18 +1,18 @@
 {
     "init_node": 0,
-    "n_neighbours": 99,
+    "n_neighbours": 5,
     "smooth": true,
     "center_and_reduce": true,
     "normalize": false,
     "sort_by_mean": false,
     "goodnodes": [118, 168, 261],
-    "number_of_nodes": 3,
+    "number_of_nodes": 1,
     "window_size": 84,
-    "target_size": 12,
+    "target_size": 5,
     "stride" : 1,
-    "communication_rounds" : 50,
-    "local_epochs": 200,
-    "fed_epochs": 50,
+    "communication_rounds" : 1,
+    "local_epochs": 1,
+    "fed_epochs": 1,
     "learning_rate": 0.001,
     "model_path" : "./"
 }
