+++ src/utils_fed.py
@@ -1,12 +1,11 @@
 
-from src.models import LSTMModel
 from src.utils_data import my_data_loader, createLoaders
 
 
 import pandas as pd
 import torch
 
-from src.models import train_model
+from src.utils_training import train_model
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
